# Conflicts:
#	.gitignore
#	bin/zoom_sdk_dotnet_wrap.dll
#	zoom_sdk_c_sharp_wrap/meeting_audio_dotnet_wrap.cpp

+++ zoom_sdk_c_sharp_wrap/meeting_audio_dotnet_wrap.cpp
@@ -16,7 +16,7 @@ namespace ZOOM_SDK_DOTNET_WRAP {
 		virtual unsigned int GetUserId()
 		{
 			if (m_pStatus)
-				return m_pStatus->GetUserId();
+				return (unsigned int)m_pStatus->GetUserId();
 			return 0;
 		}
